refactor: moved the entire component
