+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Icon/ColorToggleGroup/ColorToggleGroup.spec.tsx
@@ -2,18 +2,18 @@ import { render, fireEvent, waitFor } from '@testing-library/react'
 import { MockedProvider } from '@apollo/client/testing'
 import { EditorProvider, TreeBlock } from '@core/journeys/ui'
 import {
-  IconSize,
+  IconColor,
   IconName
-} from '../../../../../../../../../__generated__/globalTypes'
-import { IconFields } from '../../../../../../../../../__generated__/IconFields'
-import { ICON_BLOCK_SIZE_UPDATE } from './SizeToggleGroup'
-import { SizeToggleGroup } from '.'
+} from '../../../../../../../__generated__/globalTypes'
+import { IconFields } from '../../../../../../../__generated__/IconFields'
+import { ICON_BLOCK_COLOR_UPDATE } from './ColorToggleGroup'
+import { ColorToggleGroup } from '.'
 
-describe('SizeToggleGroup', () => {
-  it('should change the icon size', async () => {
+describe('ColorToggleGroup', () => {
+  it('should change the icon color', async () => {
     const icon: TreeBlock<IconFields> = {
       id: 'icon-id',
-      parentBlockId: null,
+      parentBlockId: 'buttonBlockId',
       parentOrder: null,
       __typename: 'IconBlock',
       iconName: IconName.ArrowForwardRounded,
@@ -24,13 +24,13 @@ describe('SizeToggleGroup', () => {
 
     const result = jest.fn(() => ({
       data: {
-        iconBlockUpdate: {
+        iconBlockUodate: {
           id: 'icon-id',
           journeyId: 'journeyId',
           parentBlockId: 'buttonBlockId',
           name: IconName.ArrowForwardRounded,
-          color: null,
-          size: IconSize.sm
+          color: IconColor.secondary,
+          size: null
         }
       }
     }))
@@ -40,11 +40,11 @@ describe('SizeToggleGroup', () => {
         mocks={[
           {
             request: {
-              query: ICON_BLOCK_SIZE_UPDATE,
+              query: ICON_BLOCK_COLOR_UPDATE,
               variables: {
                 id: 'id',
                 input: {
-                  size: IconSize.sm
+                  color: IconColor.secondary
                 }
               }
             },
@@ -53,12 +53,12 @@ describe('SizeToggleGroup', () => {
         ]}
       >
         <EditorProvider>
-          <SizeToggleGroup iconBlock={icon} />
+          <ColorToggleGroup iconBlock={icon} />
         </EditorProvider>
       </MockedProvider>
     )
-    expect(getByRole('button', { name: 'Medium' })).toHaveClass('Mui-selected')
-    fireEvent.click(getByRole('button', { name: 'Small' }))
+    expect(getByRole('button', { name: 'Default' })).toHaveClass('Mui-selected')
+    fireEvent.click(getByRole('button', { name: 'Secondary' }))
     await waitFor(() => expect(() => expect(result).toHaveBeenCalled()))
   })
 })
