+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Icon/Icon.stories.tsx
@@ -1,14 +1,14 @@
 import { Story, Meta } from '@storybook/react'
 import { MockedProvider } from '@apollo/client/testing'
 import { TreeBlock, EditorProvider } from '@core/journeys/ui'
-import { Drawer } from '../../../../../Drawer'
-import { journeysAdminConfig } from '../../../../../../../libs/storybook'
-import { IconFields } from '../../../../../../../../__generated__/IconFields'
+import { Drawer } from '../../../Drawer'
+import { journeysAdminConfig } from '../../../../../libs/storybook'
+import { IconFields } from '../../../../../../__generated__/IconFields'
 import {
   IconName,
   IconColor,
   IconSize
-} from '../../../../../../../../__generated__/globalTypes'
+} from '../../../../../../__generated__/globalTypes'
 import { Icon } from '.'
 
 const IconStory = {
