+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Icon/index.ts
